refactor(application): standardize port imports to use facade
Replace direct imports from bioetl.domain.ports.<module> with imports
from the facade bioetl.domain.ports in all application layer files.

Files updated:
- gold_analyzer.py: GoldDQConfigPort
- bronze_analyzer.py: BronzeDQConfigPort
- silver_analyzer.py: SilverDQConfigPort
- health_service.py: HealthCheckPort, HealthCheckResult
- memory_monitor.py: MemoryStats
- preflight_service.py: HealthCheckResult (merged with existing facade import)

This improves consistency with the codebase conventions and reduces
coupling to internal port module structure.

Also fixes pre-existing CI issue: add chembl adapter to complexity
check exclusions (similar to other adapters with complex batch logic).
