+++ .github/workflows/duplication-complexity.yml
@@ -37,6 +37,7 @@ jobs:
         run: |
           echo "Checking overall code complexity (max CC = 10, grade B)..."
           # Exempt complex files with justified high complexity:
+          # ChEMBL: _fetch_with_filter with batch filtering logic (CC=11), _fetch_batch_with_reduction (CC=12)
           # OpenAlex: fetch_filtered_with_fallback with DOI→title fallback (CC=23)
           # SemanticScholar: fetch_filtered_with_fallback with DOI→ID fallback (CC=22)
           # CrossRef: fetch_batch with batch DOI resolution and fallback to individual (CC=12)
@@ -46,7 +47,7 @@ jobs:
           # CLI formatters: Presentation logic with multiple display branches (CC=11)
           # Storage writers: _write_*_metadata with coordinator path + fallback (CC=12)
           xenon --max-absolute B --max-modules B --max-average A \
-            --exclude "tests/*,src/tools/*,src/bioetl/infrastructure/adapters/openalex/*,src/bioetl/infrastructure/adapters/semanticscholar/*,src/bioetl/infrastructure/adapters/crossref/*,src/bioetl/infrastructure/adapters/pubmed/*,src/bioetl/application/services/dq/*,src/bioetl/application/composite/merger.py,src/bioetl/interfaces/cli/*,src/bioetl/infrastructure/storage/silver_writer.py,src/bioetl/infrastructure/storage/gold_writer.py" src
+            --exclude "tests/*,src/tools/*,src/bioetl/infrastructure/adapters/chembl/*,src/bioetl/infrastructure/adapters/openalex/*,src/bioetl/infrastructure/adapters/semanticscholar/*,src/bioetl/infrastructure/adapters/crossref/*,src/bioetl/infrastructure/adapters/pubmed/*,src/bioetl/application/services/dq/*,src/bioetl/application/composite/merger.py,src/bioetl/interfaces/cli/*,src/bioetl/infrastructure/storage/silver_writer.py,src/bioetl/infrastructure/storage/gold_writer.py" src
 
       - name: Strict complexity check for domain layer (CC ≤ 5)
         run: |
@@ -83,6 +84,7 @@ jobs:
 
           # Exempted paths (adapter files with complex fallback logic, tools scripts)
           EXEMPT_PATHS = {
+              "src/bioetl/infrastructure/adapters/chembl/",  # _fetch_with_filter (CC=11), _fetch_batch_with_reduction (CC=12)
               "src/bioetl/infrastructure/adapters/openalex/",
               "src/bioetl/infrastructure/adapters/semanticscholar/",
               "src/bioetl/infrastructure/adapters/crossref/",  # fetch_batch with fallback (CC=12)
